+++ volmdlr/primitives3D.py
@@ -396,9 +396,10 @@ class ExtrudedProfile(volmdlr.Shell3D):
                 seg3 = volmdlr.LineSegment3D(seg2.points[1], seg2.points[1]-seg1.points[1] + seg1.points[0])
                 seg4 = volmdlr.LineSegment3D(seg3.points[1], seg1.points[0])
                 edges = [seg1, seg2, seg3, seg4]
-                return [edge.generated_planeface([volmdlr.Contour3D(edges)])]
+                # return [edge.generated_planeface([volmdlr.Contour3D(edges)])]
+                return [volmdlr.PlaneFace3D.from_contours3d([volmdlr.Contour3D(edges)])]
             
-            elif edge.__class__ is volmdlr.primitives3D.OpenedRoundedLineSegments2D or edge.__class__ is volmdlr.primitives3D.ClosedRoundedLineSegments2D :
+            elif edge.__class__ is volmdlr.primitives3D.OpenedRoundedLineSegments3D or edge.__class__ is volmdlr.primitives3D.ClosedRoundedLineSegments3D :
                 faces = []
                 for element in edge.primitives :
                     faces.extend(generated_faces(extru_vec, element))
@@ -413,7 +414,7 @@ class ExtrudedProfile(volmdlr.Shell3D):
             for element in self.inner_contours3d :
                 for edge in element.edges :
                     lateral_faces.extend(generated_faces(self.extrusion_vector, edge))
-        
+
         return [lower_face]+[upper_face]+lateral_faces
 
     def _bounding_box(self):